Fix case where installed path is symlink
One can install packages from a local path via composers repositories
type "path".
Those are typically symlinks which were not resolved.
We therefore add a realpath() call, but falling back to old behavior if
it didn't resolve. That way we keep independent of realpath() as
requested by the inline comment.

Resolves: https://github.com/phpactor/phpactor/issues/2727